jockey_weightをfloatに変更

+++ db/ddl/race_horce_map.sql
@@ -8,7 +8,7 @@ CREATE TABLE race_horse_map (
     wakuban         INTEGER,
     chakujun        INTEGER,
     jockey_id       INTEGER,
-    jockey_weight   INTEGER,
+    jockey_weight   NUMERIC,
     race_time       TIME,
     weight          SMALLINT,
     agari           NUMERIC,
